third_party/blink/web_tests/http/tests/loading/text-content-type-with-binary-extension-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
frame "iframe" - BeginNavigation request to 'http://127.0.0.1:8000/loading/resources/binaryData.sitx', http method GET
frame "iframe" - DidStartNavigation
main frame - didFinishDocumentLoadForFrame
frame "iframe" - didFailProvisionalLoadWithError
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
Test for https://bugs.webkit.org/show_bug.cgi?id=17360 REGRESSION: mp4 file downloaded from server is downloaded as html.

Binary data not loaded as text. Maybe PASS.
